backend/requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
openai==1.24.0
python-dotenv==1.0.1
